Create Steps I performed
These were the steps I did. Notice that the commit done on the branch1 branch appears in both the files that were created in two different branches, after merging. So commiting puts all the stuff at one place, sticking the commit to all the new files created in the new session. Creating a file is also a change. So it is the changes that are being tracked per commit, meaning the change components will have the same piece of commit message tagline.